add toastr and sweetalert

+++ src/src/wwwroot/js/views/home/index.js
@@ -54,6 +54,7 @@ function SubmitAddEdit(form) {
             success: function (data) {
                 if (data.success) {
                     popup.dialog('close');
+                    ShowMessage(data.message);
                     dataTable.ajax.reload();
                 }
             }
@@ -64,14 +65,32 @@ function SubmitAddEdit(form) {
 }
 
 function Delete(id) {
-    $.ajax({
-        type: 'DELETE',
-        url: '/api/todo/' + id,
-        success: function (data) {
-            if (data.success) {
-                dataTable.ajax.reload();
+    swal({
+        title: "Are you sure want to Delete?",
+        text: "You will not be able to restore the file!",
+        type: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#DD6B55",
+        confirmButtonText: "Yes, delete it!",
+        closeOnConfirm: true
+    }, function () {
+        $.ajax({
+            type: 'DELETE',
+            url: '/api/todo/' + id,
+            success: function (data) {
+                if (data.success) {
+                    ShowMessage(data.message);
+                    dataTable.ajax.reload();
+                }
             }
-        }
+        });
     });
+
+
+}
+
+
+function ShowMessage(msg) {
+    toastr.success(msg);
 }
